[Seed] Prefactor: extract defaulting into dedicated file
